pyproject.toml
[project]
name = "mascate"
version = "0.1.0"
description = "Assistente de Voz Edge AI para Linux"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Renato" }]
keywords = ["voice-assistant", "edge-ai", "llm", "linux", "privacy"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "rich>=13.0",
    "textual>=0.40",
    "click>=8.0",
    "tomli>=2.0",
    "tomli-w>=1.0",
    
    # Audio
    "sounddevice>=0.4",
    "numpy>=1.26",
    "scipy>=1.12",
    "onnxruntime>=1.16",
    
    # Input (hotkey)
    "pynput>=1.7",
    
    # RAG
    "qdrant-client>=1.7",
    "sentence-transformers>=2.3",
    
    # Model Management
    "huggingface-hub>=0.20",
    "tqdm>=4.66",
    
    # Utilities
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.2",
    "mypy>=1.8",
]

stt = [
    # Whisper STT via whisper.cpp bindings
    "pywhispercpp>=1.2",
]

llm = [
    # LLM via llama.cpp - requer compilacao com CMAKE_ARGS para GPU
    # CUDA: CMAKE_ARGS="-DGGML_CUDA=on" uv pip install llama-cpp-python
    # ROCm: CMAKE_ARGS="-DGGML_HIPBLAS=on" uv pip install llama-cpp-python
    "llama-cpp-python>=0.2",
]

wake = [
    # Wake word detection - NOTA: openwakeword requer tflite-runtime
    # que so tem wheels para Python 3.8-3.11. Para Python 3.12+,
    # considere usar porcupine ou vosk como alternativa.
    # Instalar manualmente: pip install tflite-runtime (se disponivel)
    # "openwakeword>=0.6",
]

cuda = [
    # Placeholder - instalar llama-cpp-python manualmente com CUDA
]

rocm = [
    # Placeholder - instalar llama-cpp-python manualmente com ROCm
]

all = [
    "mascate[stt,llm]",
]

[project.scripts]
mascate = "mascate.interface.cli:main"

[project.urls]
Homepage = "https://github.com/renato/mascate"
Documentation = "https://github.com/renato/mascate/tree/main/docs"
Repository = "https://github.com/renato/mascate"
Issues = "https://github.com/renato/mascate/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mascate"]

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict=
]

[tool.ruff.lint.isort]
known-first-party = ["mascate"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["src/mascate"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "sounddevice.*",
    "qdrant_client.*",
    "sentence_transformers.*",
    "llama_cpp.*",
]
ignore_missing_imports = true
